+++ client/components/SelfStudyPanel/AudioButton/index.tsx
@@ -1,7 +1,6 @@
 import { useGlobalTheme } from '@/styles/GlobalThemeContext';
-import { useEffect, useState } from 'react';
+import useAudio from '@/utils/audio';
 import { HiSpeakerWave } from 'react-icons/hi2';
-import { toast } from 'react-toastify';
 
 interface AudioButtonProps {
   url: string;
@@ -23,45 +22,4 @@ export default function AudioButton({ url, disabled }: AudioButtonProps) {
       />
     </>
   );
-}
-
-function useAudio(url: string): [VoidFunction] {
-  const [playing, setPlaying] = useState<boolean>(false);
-
-  const [audio, setAudio] = useState<HTMLAudioElement | null>(null);
-
-  const toggle = () => setPlaying(!playing);
-  const stopPlaying = () => setPlaying(false);
-
-  useEffect(() => {
-    audio?.pause();
-    stopPlaying();
-  }, [url, audio]);
-
-  useEffect(() => {
-    setAudio(new Audio(url));
-  }, [url]);
-
-  useEffect(() => {
-    if (!audio) return;
-
-    (async () => {
-      try {
-        await (playing ? audio.play() : audio.pause());
-      } catch (e) {
-        toast.error('음성을 재생할 수 없습니다.');
-      }
-    })();
-  }, [audio, playing]);
-
-  useEffect(() => {
-    if (!audio) return;
-    audio.addEventListener('ended', stopPlaying);
-
-    return () => {
-      audio.removeEventListener('ended', stopPlaying);
-    };
-  });
-
-  return [toggle];
 }
